5 - {add} crawling module, {change,create} AnimalVO.java,
AnimalVO.java
int -> float으로 타입 변경 get,set 메서드 매개변수 타입도 변경
AnimalVO() 생성자 추가

+++ funny/src/model/animal/AnimalDAO.java
@@ -24,7 +24,7 @@ public class AnimalDAO {
 				pstmt.setString(4, data.getAnimal_type());
 				pstmt.setString(5, data.getAnimal_gender());
 				pstmt.setString(6, data.getAnimal_age());
-				pstmt.setInt(7, data.getAnimal_weight());
+				pstmt.setFloat(7, data.getAnimal_weight());
 				pstmt.setString(8, data.getAnimal_image());
 				pstmt.executeUpdate();
 			}
